+++ NEWS.md
@@ -1,4 +1,4 @@
-# datawizard 0.2.3.9000
+# datawizard 0.2.3.9005
 
 - New functions: